requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
flask-cors==4.0.0
marshmallow==3.21.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1
psycopg2-binary
pytest==8.2.0
